% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{cols_all}
\alias{cols_all}
\alias{cols_product_all}
\alias{cols_product_emission}
\alias{cols_product_sector}
\alias{cols_product_transition_risk}
\alias{cols_company_all}
\alias{cols_company_sector}
\alias{cols_company_emission}
\alias{cols_company_transition_risk}
\title{Columns in the latest tiltOutput}
\usage{
cols_all()

cols_product_all()

cols_product_emission()

cols_product_sector()

cols_product_transition_risk()

cols_company_all()

cols_company_sector()

cols_company_emission()

cols_company_transition_risk()
}
\value{
Character.
}
\description{
Columns in the latest tiltOutput
}
\examples{
cols_all()

cols_product_all()

cols_product_emission()

cols_product_sector()

cols_product_transition_risk()

cols_company_all()

cols_company_sector()

cols_company_emission()

cols_company_transition_risk()

}
