% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{emissions}
\alias{sector}
\alias{dictionary}
\title{Datasets}
\format{
An object of class \code{tilt_profile} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 72 rows and 3 columns.

An object of class \code{tilt_profile} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 72 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81 rows and 5 columns.
}
\usage{
emissions

sector

dictionary
}
\description{
Datasets
}
\examples{
emissions

sector

dictionary
}
\keyword{datasets}
