% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{The URL where we collaboratively define columns}
\usage{
dictionary()
}
\value{
Character
}
\description{
The URL where we collaboratively define columns
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dictionary()
\dontshow{\}) # examplesIf}
}
