% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datasets.R
\name{dictionary_url}
\alias{dictionary_url}
\title{The URL where we collaboratively define columns}
\usage{
dictionary_url()
}
\value{
Character
}
\description{
The URL where we collaboratively define columns
}
\examples{
dictionary_url()
}
