% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_app}
\alias{test_app}
\title{Test the app passing data as an argument}
\usage{
test_app(company = "/pushymatador/fake/company")
}
\arguments{
\item{company}{A path to the company data.}
}
\value{
Called for its side effect.
}
\description{
Test the app passing data as an argument
}
\examples{
if (interactive()) {
# test_app()
}
}
