% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-plot.R
\name{theme_tiltplot}
\alias{theme_tiltplot}
\title{TiltPlot theme for the graphs}
\usage{
theme_tiltplot()
}
\value{
An object of class "theme", "gg".
}
\description{
TiltPlot theme for the graphs
}
\examples{
library(ggplot2, warn.conflicts = FALSE)

ggplot(mtcars) +
  geom_histogram(aes(mpg), bins = 10) +
  theme_tiltplot()
}
