% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary_url}
\alias{dictionary_url}
\alias{dictionary_edit}
\title{Dictionary}
\usage{
dictionary_url(type = c("web", "csv"))

dictionary_edit()
}
\value{
Character
}
\description{
Dictionary
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Edit the dictionary here
dictionary_edit() |> browseURL()

# Your edits are automatically published here
dictionary_url("web") |> browseURL()

# You can read it as a "csv"
dictionary_url("csv") |>
  read.csv() |>
  head()
\dontshow{\}) # examplesIf}
}
