% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltWebTool-package.R
\docType{package}
\name{tiltWebTool-package}
\alias{tiltWebTool}
\alias{tiltWebTool-package}
\title{tiltWebTool: Web Tool}
\description{
Web Tool.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltWebTool}
  \item \url{https://2degreesinvesting.github.io/tiltWebTool/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltWebTool/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Authors:
\itemize{
  \item Anne Schoenauer \email{anne.schoenauer@2degrees-investing.org}
  \item Tilman Trompke \email{tilman@2degrees-investing.org}
  \item Kalash Singhal \email{kalash@2degrees-investing.org}
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
